requirements.txt
# Core dependencies
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
requests==2.31.0

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

# Validation
marshmallow==3.20.1

# Redis for caching
redis==5.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development
black==23.12.1
flake8==7.0.0
